added analog methods to inverter.h and inverter.cpp

+++ FinalProjectFromScratch/Inverter.h
@@ -74,7 +74,7 @@ public:
 	void setOverloadAnalogValue(uint16_t value);
 	
 	int getAcInputReadings();
-	int getBattInputReadings();
+	double getBattInputReadings();
 	int getOverloadInputReadings();
